+++ crates/error/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.2-git
+
+### Minor
+
+- Add `Error::pop()` to propagate errors through space
+- Implement `From<std::io::Error>` for `Error` with `std` feature
+- Add `std` feature
+
 ## 0.1.1
 
 ### Minor